% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertnativepipeAddin.R
\name{insertnativepipeAddin}
\alias{insertnativepipeAddin}
\alias{nativepipe}
\title{Insert |> operator}
\usage{
insertnativepipeAddin()
}
\value{
|>
}
\description{
Call this function as an addin to insert \code{ |> } at the cursor position.
}
